// frontend/app/src/app/models/User.model.ts
export class User {
    userId?: number;
    username: string;
    password: string;
    securityQuestion?: string;
    securityAnswer?: string;
    firstName: string;
    lastName: string;
    gender?: string;
    address?: string;
    contactPhone?: string;
    email: string;
    profileImage?: string;
    schoolType?: string;
    currentGrade?: number;
    
  }
  